// src/app/shared/Model/updateincident.ts
export class UpdateIncidentModel {
  incident_ID:Number;
    incident_Title: string;
    incident_Description: string;
    incident_Status_Id: Number;
    start_Date: Date;
    closed_Date: Date;
    incident_Duration: Number;
    incident_Category_Id: Number;
    incident_Priority_Id: Number;
    incident_Source: string;
    incident_Location: string;
    incident_Owner: Number;
    notes: string;
    created_By: Number;
    created_Date: Date;
    modified_By: Number;
    modified_Date: Date
  }